requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy==1.24.4
Flask
gunicorn
easyocr
spacy
pillow
opencv-python-headless
openpyxl
